Move literal to a separate file
